requirements.txt
scikit-learn
pandas
decision-tree-id3